consolidate error message growlers into config editor
